Rename TextEditor to Editor
