+++ src/components/Editor/Loadable.ts
@@ -1,0 +1,12 @@
+/**
+ * Asynchronously loads the component for Editor
+ */
+
+import { loadable } from "utils/loadable";
+
+const Editor = loadable(
+  () => import("./index"),
+  (module) => module.Editor
+);
+
+export { Editor };
