+++ src/components/Editor/useTrackTextAreaScroll.ts
